# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyDocs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/docs/my-docs": {
            "get": {
                "tags": [
                    "Meus Documentos"
                ],
                "summary": "Lista todos os documentos do usuário autenticado",
                "description": "Recupera todos os documentos pertencentes ao usuário autenticado",
                "operationId": "mydocs_index",
                "responses": {
                    "200": {
                        "description": "Documentos recuperados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os documentos existentes foram recuperados com sucesso."
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "myDocs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user": {
                                                        "type": "object",
                                                        "example": []
                                                    },
                                                    "type": {
                                                        "type": "object",
                                                        "example": []
                                                    },
                                                    "file": {
                                                        "type": "string",
                                                        "example": "http://example.com/public/api/docs/filename.ext"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Meus Documentos"
                ],
                "summary": "Cadastra um novo documento para o usuário autenticado",
                "description": "Cria um novo documento com base nos dados fornecidos",
                "operationId": "mydocs_store",
                "requestBody": {
                    "description": "Informações do documento",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "file": {
                                        "description": "Arquivo do documento",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documento cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O documento foi cadastrado com sucesso."
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "doc": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Arquivo de documento não enviado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Arquivo de documento não foi enviado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Problema ao cadastrar o documento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Houve um problema ao cadastrar o documento :("
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}